Add files via upload
Added the home page for Instructor, Reviewer, Staff, Student, and updated the welcome page to redirect to the respective home page. All home pages are currently just copies of User home page, except including the database. I added the database helper assuming it was likely we would need it in the future, but it isn't currently being used.